requirements.txt
streamlit
torch
transformers
sentence-transformers
scikit-learn
numpy
faiss-cpu
accelerate
